Update and rename bluealliance.py to tbaAPI.py
All changes from FRC-YouTube-Uploader
